Make user provider more resilient
Setting permissions for a replica set on the initial chef run may
sometimes fail, probably due to the replica set still being in an
inconsistent (initial) state - despite the cluster reporting that it’s
ready. Utilising the same retry mechanism we have for node and replica
set state checks seems to solve the problem.